requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart
ultralytics==8.1.0
opencv-python-headless
streamlit==1.31.0
numpy
Pillow
requests